[STUDY-20210221] component header, main banner 제작
