+++ src/App.css
@@ -3,7 +3,7 @@
 }
 
 .App a {
-  color: black;
+  /* color: black; */
   text-decoration: none;
 }
